requirements.txt
django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.8.0
beautifulsoup4==4.9.3
selenium==3.141.0
pytz==2021.1
requests==2.25.1
fake_useragent==0.1.11
python-dateutil==2.8.2
regex==2021.4.4
uuid==1.30
webdriver-manager==3.4.2